Move WorkspaceProxy back to multi workspace client, set formatter ruby version to 3.1
