+++ lib/seam/client_multi_workspace.rb
@@ -3,7 +3,6 @@
 require_relative "parse_options"
 require_relative "lts_version"
 require_relative "auth"
-require_relative "utils/workspaces_proxy"
 
 module Seam
   class ClientMultiWorkspace
@@ -52,5 +51,19 @@ module Seam
         method, path, config
       )
     end
+  end
+end
+
+class WorkspacesProxy
+  def initialize(workspaces)
+    @workspaces = workspaces
+  end
+
+  def list(**kwargs)
+    @workspaces.list(**kwargs)
+  end
+
+  def create(**kwargs)
+    @workspaces.create(**kwargs)
   end
 end
