--- lib/seam/utils/workspaces_proxy.rb
@@ -1,15 +1,0 @@
-module Seam
-  class WorkspacesProxy
-    def initialize(workspaces)
-      @workspaces = workspaces
-    end
-
-    def list(**kwargs)
-      @workspaces.list(**kwargs)
-    end
-
-    def create(**kwargs)
-      @workspaces.create(**kwargs)
-    end
-  end
-end
